mongodb 접속 에러 해결, RunningStory에 npc와 player 리스트 추가

+++ src/main/java/com/ohgiraffers/r_pakabe/flow/runningStory/command/application/service/RunningStoryAppService.java
@@ -2,6 +2,7 @@ package com.ohgiraffers.r_pakabe.flow.runningStory.command.application.service;
 
 import com.ohgiraffers.r_pakabe.common.error.ApplicationException;
 import com.ohgiraffers.r_pakabe.common.error.ErrorCode;
+import com.ohgiraffers.r_pakabe.flow.runningStory.command.application.dto.RunningEntityDTO;
 import com.ohgiraffers.r_pakabe.flow.runningStory.command.application.dto.RunningStoryDTO;
 import com.ohgiraffers.r_pakabe.flow.runningStory.command.application.dto.RunningStoryMapper;
 import com.ohgiraffers.r_pakabe.flow.runningStory.command.domain.model.RunningStory;
@@ -44,7 +45,7 @@ public class RunningStoryAppService {
 
     @Transactional
     public RunningStoryDTO createRunningStory(RunningStoryDTO dto) {
-        RunningStory story = domainService.getRunningStory(dto.roomNum());
+        RunningStory story = domainService.getRunningStory(dto.getRoomNum());
         if (story != null) {
             throw new ApplicationException(ErrorCode.REDUNDANT_ROOM_NUMBER);
         }
@@ -53,7 +54,23 @@ public class RunningStoryAppService {
     }
 
     @Transactional
+    public RunningStoryDTO updateRunningStory(RunningStoryDTO updateDto) {
+        RunningStory story = domainService.getRunningStory(updateDto.getRoomNum());
+        if (story == null) {
+            throw new ApplicationException(ErrorCode.NO_SUCH_ROOM);
+        }
+        RunningEntityDTO entityDTO = mapper.documentToEntityDto(story);
+        mapper.updateEntityDto(entityDTO, updateDto);
+        story = domainService.updateRunningStory(mapper.entityDtoToDocument(entityDTO));
+        return mapper.documentToDto(story);
+    }
+
+    @Transactional
     public void deleteRunningStory(Integer roomNumber) {
+        RunningStory story = domainService.getRunningStory(roomNumber);
+        if (story == null) {
+            throw new ApplicationException(ErrorCode.NO_SUCH_ROOM);
+        }
         domainService.deleteRunningStory(roomNumber);
     }
